refactor: move fetch_generation_from_dp from save module to data generation module
